+++ spending-api/models/receipt_item.go
@@ -11,14 +11,14 @@ type ReceiptItem struct {
 	UUId      uuid.UUID
 	ReceiptId int
 	Name      string
-	Price     float32
+	Price     float64
 	IsDeleted bool
 	DeletedAt time.Time
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
 
-func NewReceiptItem(receiptId int, name string, price float32) *ReceiptItem {
+func NewReceiptItem(receiptId int, name string, price float64) *ReceiptItem {
 	return &ReceiptItem{
 		UUId:      uuid.New(),
 		ReceiptId: receiptId,
